Download
Unfortunately I had to skip the map chapter which may have led up to some changes that needed to be made for the search feature. But since the search feature wasn't working I had to download a copy of the lesson file. NOTE: All the previous comments are now gone :-1:  So it's probably best to take notes afterwards...