Search and rest API
Learned about search bars and using rest API. However as before I needed to copy from the lesson files. Also later on in the course a non JQuery example will be implemented and needed to be copy so refraining from commenting at the moment. However the search feature is finally working.