Deployment in the future
Deployment in the future. In order to deploy in the future ensure that your repo will be private since it will contain sensitive information that could possibly be used.

Also some instructions in the wp-config.php file on how to check/link local and deployed website. Also a method in the functions.php to check and ignore npm sites when exporting the file.

Also updated the gitignore file to ignore any of the ai1wm-backups folder/files